[KSCOR-197] fixed some findbugs errors
git-svn-id: https://svn.kuali.org/repos/student/ks-core/branches/ks-core-dev@13259 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-common-ui/src/main/java/org/kuali/student/common/ui/client/configurable/mvc/binding/HasValueBinding.java
@@ -51,7 +51,7 @@ public class HasValueBinding extends ModelWidgetBindingSupport<HasValue> {
         } else if (value instanceof Double) {
             model.set(qPath, (Double) value);
         } else if (value instanceof Byte) {
-            model.set(qPath, (Integer) value);
+            model.set(qPath, ((Byte) value).intValue());
         } else if (value instanceof Boolean) {
             model.set(qPath, (Boolean) value);
         } else if (value instanceof Date) {
